Fixed +mixins, added support for @import
Mixins could not be reused because the selector object can have
only one parent. This is fixed by shallow cloning the mixin
selector object when constructing the parse tree. The updated
test can be used to reproduce the bug in previous versions.

The @import directive is now supported. It is implemented as a
pre-compiler directive, the referenced files get inlined before
the tokenizer runs. For the sass renderer to be able to resolve
files, the options passed to the render function must have a
locals.partial(path, options) function that returns an
unprocessed sass string. When using sass with Express this will
work out of the box.

Tests have been updated / added.